volumeのURLを直接指定せず自動的にmasterから取得する (#79)

+++ app/api/fsAccess.ts
@@ -5,20 +5,51 @@ export async function fsAssign() {
     cache: "no-store",
   });
   const fsResBody = await fsRes.json();
-  if (typeof fsResBody.fid === "string") {
-    return { fid: fsResBody.fid };
+  if (
+    typeof fsResBody.fid === "string" &&
+    typeof fsResBody.publicUrl === "string"
+  ) {
+    return { fid: fsResBody.fid, volumeUrl: fsResBody.publicUrl };
   } else {
     console.log(fsRes);
     return null;
   }
 }
-export async function fsWrite(fid: string, data: Blob) {
+async function fsLookup(fid: string) {
+  const fsRes = await fetch(
+    process.env.FS_MASTER +
+      "/dir/lookup?volumeId=" +
+      fid.slice(0, fid.indexOf(",")),
+    {
+      cache: "no-store",
+    }
+  );
+  const fsResBody = await fsRes.json();
+  if (typeof fsResBody.locations?.at(0)?.publicUrl === "string") {
+    console.log(fsResBody.locations.at(0).publicUrl)
+    return fsResBody.locations.at(0).publicUrl;
+  } else {
+    console.log(fsRes);
+    return null;
+  }
+}
+export async function fsWrite(
+  fid: string,
+  volumeUrl: string | null,
+  data: Blob
+) {
+  if (volumeUrl === null) {
+    volumeUrl = await fsLookup(fid);
+  }
+  if (volumeUrl === null) {
+    return false;
+  }
   const formData = new FormData();
   const compressedBlob = await new Response(
     data.stream().pipeThrough(new CompressionStream("gzip"))
   ).blob();
   formData.append("file", compressedBlob);
-  const fsPostRes = await fetch(process.env.FS_VOLUME + "/" + fid, {
+  const fsPostRes = await fetch(volumeUrl + "/" + fid, {
     method: "POST",
     body: formData,
     cache: "no-store",
@@ -30,8 +61,14 @@ export async function fsWrite(fid: string, data: Blob) {
     return false;
   }
 }
-export async function fsRead(fid: string) {
-  const fsRes = await fetch(process.env.FS_VOLUME + "/" + fid, {
+export async function fsRead(fid: string, volumeUrl: string | null) {
+  if (volumeUrl === null) {
+    volumeUrl = await fsLookup(fid);
+  }
+  if (volumeUrl === null) {
+    return null;
+  }
+  const fsRes = await fetch(volumeUrl + "/" + fid, {
     cache: "no-store",
   });
   if (fsRes.ok && fsRes.body) {
@@ -44,8 +81,14 @@ export async function fsRead(fid: string) {
     return null;
   }
 }
-export async function fsDelete(fid: string) {
-  const fsRes = await fetch(process.env.FS_VOLUME + "/" + fid, {
+export async function fsDelete(fid: string, volumeUrl: string | null) {
+  if (volumeUrl === null) {
+    volumeUrl = await fsLookup(fid);
+  }
+  if (volumeUrl === null) {
+    return false;
+  }
+  const fsRes = await fetch(volumeUrl + "/" + fid, {
     method: "DELETE",
     cache: "no-store",
   });
